Moved cmdletbinding and changed a throw
Moved cmdletbinding outside of the param block (like Microsoft suggests).

In the case where the file exists and the user chooses not to overwrite the file, the script will no longer throw, but rather inform the user that this is the case and exit the script. This is because it's not an error, but rather a user choice.